requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.1.24
fonttools==4.55.8
google-auth==2.38.0
google-genai==1.2.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.20
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.0a2
pydantic_core==2.29.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
zipp==3.21.0
